tests/submission/Infraware/Forms/contents/Forms/button_formmethod.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>button_formMethod</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="get"
		  enctype=""
		  action=""
		  name="input_form">
	<p><button id='button_id'>button</button></p>
	</form>

	<script>
		
		var button = document.getElementById("button_id");
		button.formMethod = "post";

		if (typeof(button.formMethod) == "string") {
			test(function() {
				assert_equals(button.formMethod, "post", "formMethod attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("formMethod attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
